move type check into a typecheck mod
